// src/models/user-service-add-human-user-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Zitadel SDK
 * The Zitadel SDK is a convenience wrapper around the Zitadel APIs to assist you in integrating with your Zitadel environment. This SDK enables you to handle resources, settings, and configurations within the Zitadel platform.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime.js';
import type { UserServiceDetails } from './user-service-details.js';
import {
    UserServiceDetailsFromJSON,
    UserServiceDetailsFromJSONTyped,
    UserServiceDetailsToJSON,
    UserServiceDetailsToJSONTyped,
} from './user-service-details.js';

/**
 * 
 * @export
 * @interface UserServiceAddHumanUserResponse
 */
export interface UserServiceAddHumanUserResponse {
    /**
     * 
     * @type {string}
     * @memberof UserServiceAddHumanUserResponse
     */
    userId?: string;
    /**
     * 
     * @type {UserServiceDetails}
     * @memberof UserServiceAddHumanUserResponse
     */
    details?: UserServiceDetails;
    /**
     * 
     * @type {string}
     * @memberof UserServiceAddHumanUserResponse
     */
    emailCode?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UserServiceAddHumanUserResponse
     */
    phoneCode?: string | null;
}

/**
 * Check if a given object implements the UserServiceAddHumanUserResponse interface.
 */
export function instanceOfUserServiceAddHumanUserResponse(value: object): value is UserServiceAddHumanUserResponse {
    return true;
}

export function UserServiceAddHumanUserResponseFromJSON(json: any): UserServiceAddHumanUserResponse {
    return UserServiceAddHumanUserResponseFromJSONTyped(json, false);
}

export function UserServiceAddHumanUserResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): UserServiceAddHumanUserResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'userId': json['userId'] == null ? undefined : json['userId'],
        'details': json['details'] == null ? undefined : UserServiceDetailsFromJSON(json['details']),
        'emailCode': json['emailCode'] == null ? undefined : json['emailCode'],
        'phoneCode': json['phoneCode'] == null ? undefined : json['phoneCode'],
    };
}

export function UserServiceAddHumanUserResponseToJSON(json: any): UserServiceAddHumanUserResponse {
    return UserServiceAddHumanUserResponseToJSONTyped(json, false);
}

export function UserServiceAddHumanUserResponseToJSONTyped(value?: UserServiceAddHumanUserResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'userId': value['userId'],
        'details': UserServiceDetailsToJSON(value['details']),
        'emailCode': value['emailCode'],
        'phoneCode': value['phoneCode'],
    };
}